backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Environment and utilities
python-dotenv==1.0.0
requests==2.31.0

# AWS SDK for Amazon Polly
boto3==1.34.0

# Audio processing (optional, for advanced metrics)
pydub==0.25.1

# Statistics and data processing
numpy==1.26.0
